finish the change and test
