+++ examples/cpp/inference/LLAMA/dataloader.cu
@@ -26,9 +26,14 @@ void DataLoader::load_input(Task const *task,
   assert(task->regions.size() == 2);
   //   SampleIdxs *meta = (SampleIdxs *)task->local_args;
 
+  // DataLoaderNextBatchInput const input_struct =
+  //     *((DataLoaderNextBatchInput *)task->args);
+  // BatchConfig::SampleIdxs const &meta = input_struct.meta;
+
   DataLoaderNextBatchInput const input_struct =
       *((DataLoaderNextBatchInput *)task->args);
-  BatchConfig::SampleIdxs const &meta = input_struct.meta;
+  BatchConfig *bc = input_struct.bc;
+
   std::map<size_t, long> const &prev_batch_preds =
       input_struct.prev_batch_preds;
 
@@ -50,38 +55,30 @@ void DataLoader::load_input(Task const *task,
   coord_t batch_size =
       batch_input_domain.hi()[1] - batch_input_domain.lo()[1] + 1;
 
-  // copy 1 token from each batch
-  //  FIXME: currently assume continous indices
-  size_t guid = meta.guids[0];
-  size_t start_idx = meta.token_indexes[0].token_position;
+  size_t guid = bc->tokensInfo[0].guid;
+  size_t start_idx = bc->tokensInfo[0].abs_depth_in_request;
   size_t dst_idx = 0;
 
-  std::cout << "num samples " << meta.num_samples << "\n";
-
-  for (size_t i = 0; i <= meta.num_samples; i++) {
-
-    // if the first token in one request
-    if (i == meta.num_samples || meta.guids[i] != guid) {
+  for (int i = 0; i <= bc->num_active_tokens(); i++) {
+    if (i == bc->num_active_tokens() || bc->tokensInfo[i].guid != guid) {
       size_t tokens_to_copy =
-          (meta.token_indexes[i - 1].token_position - start_idx + 1);
-      std::cout << "size to copy:  " << tokens_to_copy << "\n";
+          (bc->tokensInfo[i - 1].abs_depth_in_request - start_idx + 1);
+      
 
-      if (tokens_to_copy > 1 || meta.token_indexes[i - 1].token_position <
-                                    meta.token_indexes[i - 1].initial_length) {
+      size_t request_index = bc->tokensInfo[i - 1].request_index;
+      size_t token_start_offset = bc->requestsInfo[request_index].token_start_offset;
+    
+      std::cout << "size to copy:  " << tokens_to_copy << ", start offset: " << token_start_offset << "\n";
+      if (tokens_to_copy > 1 || token_start_offset == 0) {
         // token pos < init length, the init length is the input sentence length
         // so this is the initial input, load from file.
-
         size_t copy_start_index = guid * llamaconfig.sentence_len;
         std::cout << "copy index:  " << copy_start_index << "\n";
         copy_kernel<<<GET_BLOCKS(tokens_to_copy), CUDA_NUM_THREADS>>>(
             batch_input.ptr + dst_idx,
             full_input.ptr + copy_start_index,
             tokens_to_copy);
-
         std::cout << "------------req---------------: " << guid << "\n";
-        if (guid == 0) {
-          std::cout << "guid: " << meta.guids[i] << ", i: " << i << std::endl;
-        }
         for (int i = 0; i < 8; i++) {
           std::cout << "value: " << full_input.ptr[copy_start_index + i]
                     << std::endl;
@@ -92,21 +89,30 @@ void DataLoader::load_input(Task const *task,
         // for token by token generating, get token from the previous inference.
 
         long token = prev_batch_preds.at(guid);
-        std::cout << "next iter  " << meta.token_indexes[i - 1].token_position
+        // std::cout << "next iter  " << meta.token_indexes[i -
+        // 1].token_position
+        //           << ", dst_idx: " << dst_idx << ", token:" << token << "\n";
+
+        std::cout << "next iter  " << bc->tokensInfo[i - 1].abs_depth_in_request
                   << ", dst_idx: " << dst_idx << ", token:" << token << "\n";
         long *dst_ptr = batch_input.ptr + dst_idx;
 
         cudaMemcpy(dst_ptr, &token, sizeof(long), cudaMemcpyHostToDevice);
       }
 
-      // update for next req
-      if (i < meta.num_samples) {
-        guid = meta.guids[i];
-        start_idx = meta.token_indexes[i].token_position;
+      if (i < bc->num_active_tokens()) {
+        guid = bc->tokensInfo[i].guid;
+        start_idx = bc->tokensInfo[i].abs_depth_in_request;
       }
       dst_idx = i;
     }
   }
+
+  // copy 1 token from each batch
+  //  FIXME: currently assume continous indices
+  // size_t guid = meta.guids[0];
+  // size_t start_idx = meta.token_indexes[0].token_position;
+  // size_t dst_idx = 0;
 
   std::cout << "load input finished....." << std::endl;
 }
