+++ src/ops/inc_multihead_self_attention.cu
@@ -42,7 +42,7 @@ __global__ void build_w_out_tensor(float const *weight_ptr,
 
 __global__ void apply_rotary_embedding(float *input_ptr,
                                        cuFloatComplex *complex_input,
-                                       BatchConfig::token_idxs const *id_map,
+                                       BatchConfig::PerTokenInfo const *tokenInfos,
                                        int qProjSize,
                                        int kProjSize,
                                        int num_heads,
@@ -72,7 +72,8 @@ __global__ void apply_rotary_embedding(float *input_ptr,
     //  int head_idx = i / (num_tokens * proj_size);
     int token_idx =
         (i - head_idx * (num_tokens * proj_size / 2)) / (proj_size / 2);
-    size_t pos = id_map[token_idx].token_position;
+    // size_t pos = id_map[token_idx].token_position;
+    size_t pos = tokenInfos[token_idx].abs_depth_in_request;
 
     // float before_real = complex_input[i].x, before_complex =
     // complex_input[i].y;
@@ -225,7 +226,7 @@ void inference_kernel1(IncMultiHeadSelfAttentionMeta const *m,
                              0,
                              stream>>>(output_ptr,
                                        complex_input,
-                                       m->dev_token2ids,
+                                       m->token_infos,
                                        m->qProjSize,
                                        m->kProjSize,
                                        m->num_heads,
@@ -240,7 +241,7 @@ void inference_kernel1(IncMultiHeadSelfAttentionMeta const *m,
                              0,
                              stream>>>(output_ptr,
                                        complex_input,
-                                       m->dev_token2ids,
+                                       m->token_infos,
                                        m->qProjSize,
                                        m->kProjSize,
                                        m->num_heads,
@@ -254,7 +255,7 @@ void inference_kernel1(IncMultiHeadSelfAttentionMeta const *m,
 
 __global__ void store_kv_cache(float const *devQKVProjArray,
                                float *cache_ptr,
-                               BatchConfig::token_idxs const *id_map,
+                               BatchConfig::PerTokenInfo const *tokenInfos,
                                int qProjSize,
                                int kProjSize,
                                int vProjSize,
@@ -275,8 +276,10 @@ __global__ void store_kv_cache(float const *devQKVProjArray,
     float val =
         devQKVProjArray[head_idx * qkv_block_size + current_head_block_size +
                         token_idx * proj_size + data_idx];
-    int const req_id = id_map[token_idx].request_index;
-    int const tok_id = id_map[token_idx].token_position;
+    // int const req_id = id_map[token_idx].request_index;
+    // int const tok_id = id_map[token_idx].token_position;
+    int const req_id = tokenInfos[token_idx].request_index;
+    int const tok_id = tokenInfos[token_idx].abs_depth_in_request;
 
     cache_ptr[req_id * (num_heads * max_seq_len * proj_size) +
               head_idx * (max_seq_len * proj_size) + tok_id * proj_size +
@@ -295,7 +298,7 @@ void inference_kernel2(IncMultiHeadSelfAttentionMeta const *m,
                      0,
                      stream>>>(m->devQKVProjArray,
                                m->keyCache,
-                               m->dev_token2ids,
+                               m->token_infos,
                                m->qProjSize,
                                m->kProjSize,
                                m->vProjSize,
@@ -310,7 +313,7 @@ void inference_kernel2(IncMultiHeadSelfAttentionMeta const *m,
                      0,
                      stream>>>(m->devQKVProjArray,
                                m->valueCache,
-                               m->dev_token2ids,
+                               m->token_infos,
                                m->qProjSize,
                                m->kProjSize,
                                m->vProjSize,
@@ -366,8 +369,9 @@ void inference_kernel3(IncMultiHeadSelfAttentionMeta const *m,
     if (bc->request_completed[i]) {
       continue;
     }
-    int num_new_tokens = bc->num_processing_tokens[i];
-    int total_tokens = bc->token_last_available_idx[i] + 1;
+    int num_new_tokens = bc->requestsInfo[i].num_tokens_in_batch;
+    int total_tokens = bc->requestsInfo[i].token_start_offset + bc->requestsInfo[i].num_tokens_in_batch;
+    // bc->token_last_available_idx[i] + 1;
     // Compute (QK^T/sqrt(d_k))
     int m_ = num_new_tokens;
     int n = total_tokens;
@@ -583,9 +587,9 @@ void IncMultiHeadSelfAttention::inference_kernel_wrapper(
   inference_kernel1(m, bc, input_ptr, weight_ptr, m->devQKVProjArray, stream);
 
   // phase 2: Update key/val cache
-  cudaMemcpyAsync(m->dev_token2ids,
-                  &(bc->token2ids.token_indexes),
-                  bc->MAX_NUM_TOKENS * sizeof(BatchConfig::token_idxs),
+  cudaMemcpyAsync(m->token_infos,
+                  &(bc->tokensInfo),
+                  bc->MAX_NUM_TOKENS * sizeof(BatchConfig::tokensInfo),
                   cudaMemcpyHostToDevice,
                   stream);
 
@@ -662,7 +666,7 @@ IncMultiHeadSelfAttentionMeta::IncMultiHeadSelfAttentionMeta(
         num_heads * kProjSize * BatchConfig::MAX_NUM_REQUESTS * MAX_SEQ_LEN;
     size_t value_cache_size =
         num_heads * vProjSize * BatchConfig::MAX_NUM_REQUESTS * MAX_SEQ_LEN;
-    size_t token2ids_size = BatchConfig::MAX_NUM_TOKENS;
+    size_t tokeninfo_size = BatchConfig::MAX_NUM_TOKENS;
     size_t qk_prod_size =
         BatchConfig::MAX_NUM_TOKENS * BatchConfig::MAX_NUM_TOKENS * num_heads;
     size_t attn_heads_size =
@@ -673,8 +677,8 @@ IncMultiHeadSelfAttentionMeta::IncMultiHeadSelfAttentionMeta(
         (qkv_max_proj_size + key_cache_size + value_cache_size +
          2 * qk_prod_size + attn_heads_size + W_out_contiguous_size) *
             sizeof(float) +
-        token2ids_size *
-            sizeof(BatchConfig::token_idxs); // more components will
+        tokeninfo_size *
+            sizeof(BatchConfig::PerTokenInfo); // more components will
                                              // be added here later
 
     Realm::Rect<1, coord_t> bounds(Realm::Point<1, coord_t>(0),
@@ -691,8 +695,8 @@ IncMultiHeadSelfAttentionMeta::IncMultiHeadSelfAttentionMeta(
     devQKVProjArray = (float *)reserveInst.pointer_untyped(0, sizeof(char));
     keyCache = (float *)devQKVProjArray + qkv_max_proj_size;
     valueCache = (float *)keyCache + key_cache_size;
-    dev_token2ids = (BatchConfig::token_idxs *)(valueCache + value_cache_size);
-    qk_prods = (float *)(dev_token2ids + token2ids_size);
+    token_infos = (BatchConfig::PerTokenInfo *)(valueCache + value_cache_size);
+    qk_prods = (float *)(token_infos + tokeninfo_size);
     qk_prods_softmax = (float *)(qk_prods + qk_prod_size);
     attn_heads = (float *)qk_prods_softmax + qk_prod_size;
     W_out_contiguous = (float *)attn_heads + attn_heads_size;
